test(db): extract document mutator tests
